Rename conv.hh -> conv.h
